requirements.txt
flask==2.3.2
google-generativeai==0.5.4
flask-cors==4.0.0
gunicorn==22.0.0
werkzeug==2.3.6
mysql-connector-python==8.0.33